requirements.txt
Django>=4.3
djangorestframework
djangorestframework-simplejwt
django-tenants
drf-yasg
psycopg2-binary
